requirements.txt
requests
python-logging-loki
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
psutil